requirements.txt
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
airtest==1.2.4
anyio==3.5.0
appdirs==1.4.4
Appium-Python-Client==2.1.2
APScheduler==3.7.0
asgiref==3.3.4
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brotli==1.0.9
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
decorator==5.1.1
Deprecated==1.2.13
Django==2.2.25
django-apscheduler==0.6.0
django-cors-headers==3.7.0
djangorestframework==3.13.1
elasticsearch==7.16.3
environs==9.4.0
facebook-wda==1.4.6
Flask==1.1.4
frida==15.1.14
frida-tools==10.4.1
frozenlist==1.2.0
furl==2.1.3
gerapy==0.9.10
gerapy-prometheus-exporter==0.1.0
gerapy-selenium==0.0.3
gevent==21.12.0
greenlet==1.1.2
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hrpc==1.0.9
httpcore==0.14.4
httpx==0.21.3
httpx-socks==0.7.2
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.0
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kaitaistruct==0.8
ldap3==2.7
loguru==0.5.3
lxml==4.7.1
MarkupSafe==2.0.1
marshmallow==3.14.1
mitmdump==1.1.2
mitmproxy==5.2
motor==2.5.1
msgpack==1.0.3
mss==6.1.0
multidict==5.2.0
numpy==1.19.3
olefile==0.46
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
orderedmultidict==1.0.1
outcome==1.1.0
packaging==21.3
parsel==1.6.0
passlib==1.7.4
pika==1.2.0
Pillow==9.0.0
playwright==1.17.2
pocoui==1.0.85
priority==1.3.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
Protego==0.1.16
protobuf==3.11.3
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pydivert==2.1.0
pyee==8.2.2
PyExecJS==1.5.1
Pygments==2.11.2
pymongo==3.12.3
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pypiwin32==223
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dotenv==0.19.2
python-scrapyd-api==2.1.2
python-socks==2.0.2
pytz==2021.3
pywasm==1.0.7
pywin32==303
pywinauto==0.6.3
queuelib==1.6.2
redis==4.1.0
requests==2.27.1
retry==0.9.2
retrying==1.3.3
rfc3986==1.5.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.6
Scrapy==2.5.1
scrapy-redis==0.7.2
Scrapy-Redis-BloomFilter==0.8.1
scrapy-splash==0.8.0
scrapyd==1.3.0
scrapyd-client==1.2.0
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.1.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typing_extensions==4.0.1
tzlocal==2.1
urllib3==1.26.8
urwid==2.0.1
w3lib==1.22.0
wasmer==1.1.0
wasmer_compiler_cranelift==1.1.0
wcwidth==0.2.5
websocket==0.2.1
websocket-client==0.48.0
websockets==10.1
Werkzeug==1.0.1
win32-setctime==1.0.4
wincertstore==0.2
wrapt==1.13.3
wsproto==0.15.0
yarl==1.7.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zstandard==0.13.0
